Extract `Http4sMUnitTestCreator` to its own file
